+++ src/lilmachine/io/input/BlockingQueueMultipleInputHandler.java
@@ -6,18 +6,18 @@ import java.util.concurrent.LinkedBlockingQueue;
 
 public class BlockingQueueMultipleInputHandler implements InputHandler{
 
-    private final Queue<Integer> queue;
+    private final Queue<Long> queue;
 
     public BlockingQueueMultipleInputHandler(){
         queue = new LinkedBlockingQueue<>();
     }
 
-    public void addInput(int input){
+    public void addInput(long input){
         queue.add(input);
     }
 
     @Override
-    public int getInput() {
+    public long getInput() {
         while(queue.isEmpty()) {}
         return queue.poll();
     }
