+++ src/lilmachine/io/input/QueueBasedMultipleInputHandler.java
@@ -6,18 +6,18 @@ import java.util.concurrent.LinkedBlockingQueue;
 
 public class QueueBasedMultipleInputHandler implements InputHandler{
 
-    private final Queue<Integer> queue;
+    private final Queue<Long> queue;
 
     public QueueBasedMultipleInputHandler(){
         queue = new LinkedBlockingQueue<>();
     }
 
-    public void addInput(int input){
+    public void addInput(long input){
         queue.add(input);
     }
 
     @Override
-    public int getInput() {
-        return Objects.requireNonNullElse(queue.poll(), 0);
+    public long getInput() {
+        return Objects.requireNonNullElse(queue.poll(), 0L);
     }
 }
