+++ src/lilmachine/io/input/SystemInputHandler.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 
 public class SystemInputHandler implements InputHandler{
     @Override
-    public int getInput() {
+    public long getInput() {
         Scanner reader = new Scanner(System.in);
         System.out.print("> ");
         int input = reader.nextInt();
