+++ src/lilmachine/io/output/OutputHandler.java
@@ -1,5 +1,5 @@
 package lilmachine.io.output;
 
 public interface OutputHandler {
-    void handleOutput(int output);
+    void handleOutput(long output);
 }
